mpc setup revert trusted setup
